// api/network/requests.ts
export namespace LoginRequest
{
    export interface Query
    {
        url: string
        skin: string
        i: string
    }
    export interface FormData
    {
        url: string
        timezone: string
        skin: string
        user2: string
        user: string
        password: string
    }
}